// frontend/plugins/welcome/src/api/models/EntEquipmentrental.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SUT SA Example API Patient
 * This is a sample server for SUT SE 2563
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@swagger.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    EntEquipmentrentalEdges,
    EntEquipmentrentalEdgesFromJSON,
    EntEquipmentrentalEdgesFromJSONTyped,
    EntEquipmentrentalEdgesToJSON,
} from './';

/**
 * 
 * @export
 * @interface EntEquipmentrental
 */
export interface EntEquipmentrental {
    /**
     * RENTALAMOUNT holds the value of the "RENTALAMOUNT" field.
     * @type {number}
     * @memberof EntEquipmentrental
     */
    rENTALAMOUNT?: number;
    /**
     * RENTALDATE holds the value of the "RENTALDATE" field.
     * @type {string}
     * @memberof EntEquipmentrental
     */
    rENTALDATE?: string;
    /**
     * RETURNDATE holds the value of the "RETURNDATE" field.
     * @type {string}
     * @memberof EntEquipmentrental
     */
    rETURNDATE?: string;
    /**
     * 
     * @type {EntEquipmentrentalEdges}
     * @memberof EntEquipmentrental
     */
    edges?: EntEquipmentrentalEdges;
    /**
     * ID of the ent.
     * @type {number}
     * @memberof EntEquipmentrental
     */
    id?: number;
}

export function EntEquipmentrentalFromJSON(json: any): EntEquipmentrental {
    return EntEquipmentrentalFromJSONTyped(json, false);
}

export function EntEquipmentrentalFromJSONTyped(json: any, ignoreDiscriminator: boolean): EntEquipmentrental {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'rENTALAMOUNT': !exists(json, 'RENTALAMOUNT') ? undefined : json['RENTALAMOUNT'],
        'rENTALDATE': !exists(json, 'RENTALDATE') ? undefined : json['RENTALDATE'],
        'rETURNDATE': !exists(json, 'RETURNDATE') ? undefined : json['RETURNDATE'],
        'edges': !exists(json, 'edges') ? undefined : EntEquipmentrentalEdgesFromJSON(json['edges']),
        'id': !exists(json, 'id') ? undefined : json['id'],
    };
}

export function EntEquipmentrentalToJSON(value?: EntEquipmentrental | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'RENTALAMOUNT': value.rENTALAMOUNT,
        'RENTALDATE': value.rENTALDATE,
        'RETURNDATE': value.rETURNDATE,
        'edges': EntEquipmentrentalEdgesToJSON(value.edges),
        'id': value.id,
    };
}